+++ driver-hd44780/CHANGELOG.md
@@ -4,3 +4,4 @@
 - initial version
 ## [0.2] - 2017-12-28
 - changed package name and group id
+- improved implementation of delayMicroseconds()